pyproject.toml
[project]
name = "minutes-of-meeting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "chroma>=0.2.0",
    "chromadb>=1.5.8",
    "edge-tts>=7.2.8",
    "fastapi>=0.136.1",
    "faster-whisper>=1.2.1",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.2",
    "langchain-ollama>=1.1.0",
    "langchain-text-splitters>=1.1.2",
    "numpy>=2.4.4",
    "ollama>=0.6.1",
    "python-multipart>=0.0.26",
    "python-pptx>=1.0.2",
    "requests>=2.33.1",
    "sentence-transformers>=5.4.1",
    "streamlit>=1.56.0",
    "uvicorn>=0.46.0",
]
